Change parser name and add Argparser
